refactor: move test logic to origlang-testsuite
