+++ package/origlang-testsuite/Cargo.toml
@@ -6,3 +6,10 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+log = "0.4.20"
+origlang-ast = { version = "0.1.0", path = "../origlang-ast" }
+origlang-compiler = { version = "0.1.0", path = "../origlang-compiler" }
+origlang-ir = { version = "0.1.0", path = "../origlang-ir" }
+origlang-ir-optimizer = { version = "0.1.0", path = "../origlang-ir-optimizer" }
+origlang-runtime = { version = "0.1.0", path = "../origlang-runtime" }
+thiserror = "1.0.49"
